Module algobound::params

Enums

Mode

An enum specifying if the current simulation is to yield a lower bound or an upper bound to the true lambda.

Constants

ADV_COINS

The number of coins drawn for the adversary assuming their stake is spread across arbitrarily many accounts.

ALPHA

The fraction of stake controlled by the adversary.

BETA

The fraction of honest stake credential broadcasts which the adversary is permitted to see before having to decide what to broadcast each round.

CHERNOFF_ERROR

Upper bound on the probability that a single inflate / deflate procedure fails to produce a stochastic upper / lower bound.

EPSILON

Discretization interval size for all functions of the adversary’s rewards. Namely, used in Pdf, Cdf, Emax, and the second argument to Table.

ETA

Discretization interval size for the first argument to Table.

MODE

The type of bound to be computed.

PARALLELISM_FACTOR

The number of threads spawned for concurrent portions of the code.

ROUND_DEPTH

The number of rounds simulated before returning the adversary’s reward for a given value of lambda.

SAMPLES_DRAWN

The number of samples drawn to compute the reward distribution on each round.